+++ x/xibc/core/packet/types/packet.go
@@ -130,21 +130,13 @@ func (p Packet) ValidateBasic() error {
 }
 
 // NewResultAcknowledgement returns a new instance of Acknowledgement using an Acknowledgement_Result type in the Response field.
-func NewResultAcknowledgement(code uint64, results []byte, message, relayer string) Acknowledgement {
+func NewAcknowledgement(code uint64, results []byte, message, relayer string, feeOption uint64) Acknowledgement {
 	return Acknowledgement{
-		Code:    code,
-		Result:  results,
-		Message: message,
-		Relayer: relayer,
-	}
-}
-
-// NewErrorAcknowledgement returns a new instance of Acknowledgement using an Acknowledgement_Error type in the Response field.
-func NewErrorAcknowledgement(code uint64, message, relayer string) Acknowledgement {
-	return Acknowledgement{
-		Code:    code,
-		Message: message,
-		Relayer: relayer,
+		Code:      code,
+		Result:    results,
+		Message:   message,
+		Relayer:   relayer,
+		FeeOption: feeOption,
 	}
 }
 
@@ -205,35 +197,6 @@ func (result *Result) ABIDecode(bz []byte) error {
 		return err
 	}
 	return json.Unmarshal(bzTmp, &result)
-}
-
-type WPacket struct {
-	SrcChain string
-	// packet base data
-	DestChain string
-	Sequence  uint64
-	Sender    string
-	// transfer data. keep empty if not used.
-	TransferData []byte
-	// call data. keep empty if not used
-	CallData []byte
-	// callback data
-	CallbackAddress string
-	// fee option
-	FeeOption uint64
-}
-
-func (p Packet) ToWPacket() WPacket {
-	return WPacket{
-		p.SourceChain,
-		p.DestinationChain,
-		p.Sequence,
-		p.Sender,
-		p.TransferData,
-		p.CallData,
-		p.CallbackAddress,
-		p.FeeOption,
-	}
 }
 
 func (e *EventSendPacket) ABIDecode(bz []byte) error {
