+++ x/xibc/genesis_test.go
@@ -2,9 +2,6 @@ package xibc_test
 
 import (
 	"fmt"
-	"testing"
-
-	"github.com/stretchr/testify/suite"
 
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 
@@ -18,37 +15,6 @@ import (
 	xibctesting "github.com/teleport-network/teleport/x/xibc/testing"
 	"github.com/teleport-network/teleport/x/xibc/types"
 )
-
-const (
-	chainName  = "tendermint-0"
-	chainName2 = "tendermin-1"
-)
-
-var clientHeight = clienttypes.NewHeight(0, 10)
-
-type XIBCTestSuite struct {
-	suite.Suite
-
-	coordinator *xibctesting.Coordinator
-
-	chainA *xibctesting.TestChain
-	chainB *xibctesting.TestChain
-	chainC *xibctesting.TestChain
-}
-
-// SetupTest creates a coordinator with 2 test chains.
-func (suite *XIBCTestSuite) SetupTest() {
-	suite.coordinator = xibctesting.NewCoordinator(suite.T(), 3)
-
-	suite.chainA = suite.coordinator.GetChain(xibctesting.GetChainID(0))
-	suite.chainB = suite.coordinator.GetChain(xibctesting.GetChainID(1))
-	suite.chainC = suite.coordinator.GetChain(xibctesting.GetChainID(2))
-
-}
-
-func TestXIBCTestSuite(t *testing.T) {
-	suite.Run(t, new(XIBCTestSuite))
-}
 
 func (suite *XIBCTestSuite) TestValidateGenesis() {
 	header := suite.chainA.CreateTMClientHeader(
