+++ x/xibc/integration_test.go
@@ -2,14 +2,12 @@ package xibc_test
 
 import (
 	"encoding/hex"
-	"fmt"
 	"math/big"
 	"strconv"
 	"strings"
+	"testing"
 
-	"github.com/teleport-network/teleport/syscontracts/agent"
-
-	"github.com/teleport-network/teleport/syscontracts"
+	"github.com/stretchr/testify/suite"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
@@ -25,9 +23,41 @@ import (
 	crosschaincontract "github.com/teleport-network/teleport/syscontracts/xibc_crosschain"
 	packetcontract "github.com/teleport-network/teleport/syscontracts/xibc_packet"
 	aggregatetypes "github.com/teleport-network/teleport/x/aggregate/types"
+	clienttypes "github.com/teleport-network/teleport/x/xibc/core/client/types"
 	packettypes "github.com/teleport-network/teleport/x/xibc/core/packet/types"
 	xibctesting "github.com/teleport-network/teleport/x/xibc/testing"
 )
+
+const (
+	chainName  = "tendermint-0"
+	chainName2 = "tendermin-1"
+)
+
+var clientHeight = clienttypes.NewHeight(0, 10)
+
+type XIBCTestSuite struct {
+	suite.Suite
+
+	coordinator *xibctesting.Coordinator
+
+	chainA *xibctesting.TestChain
+	chainB *xibctesting.TestChain
+	chainC *xibctesting.TestChain
+}
+
+// SetupTest creates a coordinator with 2 test chains.
+func (suite *XIBCTestSuite) SetupTest() {
+	suite.coordinator = xibctesting.NewCoordinator(suite.T(), 3)
+
+	suite.chainA = suite.coordinator.GetChain(xibctesting.GetChainID(0))
+	suite.chainB = suite.coordinator.GetChain(xibctesting.GetChainID(1))
+	suite.chainC = suite.coordinator.GetChain(xibctesting.GetChainID(2))
+
+}
+
+func TestXIBCTestSuite(t *testing.T) {
+	suite.Run(t, new(XIBCTestSuite))
+}
 
 func (suite *XIBCTestSuite) TestBindToken() {
 	// deploy ERC20
@@ -143,11 +173,12 @@ func (suite *XIBCTestSuite) TestCrossChainTransferERC20() {
 		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
 		FeeOption:        0,
 	}
-	ack := packettypes.NewResultAcknowledgement(
+	ack := packettypes.NewAcknowledgement(
 		0,
 		[]byte(""),
 		"",
 		strings.ToLower(suite.chainB.SenderAcc.String()),
+		0,
 	)
 	ackData, err := ack.ABIPack()
 	suite.Require().NoError(err)
@@ -176,355 +207,358 @@ func (suite *XIBCTestSuite) TestCrossChainTransferERC20() {
 	suite.Require().Equal(suite.chainB.SenderAcc.String(), ack.Relayer)
 }
 
-func (suite *XIBCTestSuite) TestCrossChainTransferBaseAndTransferBack() {
-	suite.SetupTest()
+// func (suite *XIBCTestSuite) TestCrossChainTransferBaseAndTransferBack() {
+// 	suite.SetupTest()
 
-	// setup testing conditions
-	pathAToB := xibctesting.NewPath(suite.chainA, suite.chainB)
-	suite.coordinator.SetupClients(pathAToB)
+// 	// setup testing conditions
+// 	pathAToB := xibctesting.NewPath(suite.chainA, suite.chainB)
+// 	suite.coordinator.SetupClients(pathAToB)
 
-	// deploy ERC20
-	chainABase := common.HexToAddress("0x0000000000000000000000000000000000000000")
-	chainBERC20 := suite.DeployERC20ByCrossChain(suite.chainB)
+// 	// deploy ERC20
+// 	chainABase := common.HexToAddress("0x0000000000000000000000000000000000000000")
+// 	chainBERC20 := suite.DeployERC20ByCrossChain(suite.chainB)
 
-	// add erc20 trace
-	err := suite.chainB.App.AggregateKeeper.RegisterERC20Trace(
-		suite.chainB.GetContext(),
-		chainBERC20,
-		strings.ToLower(chainABase.String()),
-		suite.chainA.ChainID,
-		uint8(0),
-	)
-	suite.Require().NoError(err)
-	// check ERC20 trace
-	token, amount, exist, err := suite.chainB.App.AggregateKeeper.QueryERC20Trace(
-		suite.chainB.GetContext(),
-		chainBERC20,
-		suite.chainA.ChainID,
-	)
-	suite.Require().NoError(err)
-	suite.Require().True(exist)
-	suite.Equal(token, strings.ToLower(chainABase.String()))
-	suite.Equal(amount.Int64(), int64(0))
+// 	// add erc20 trace
+// 	err := suite.chainB.App.AggregateKeeper.RegisterERC20Trace(
+// 		suite.chainB.GetContext(),
+// 		chainBERC20,
+// 		strings.ToLower(chainABase.String()),
+// 		suite.chainA.ChainID,
+// 		uint8(0),
+// 	)
+// 	suite.Require().NoError(err)
+// 	// check ERC20 trace
+// 	token, amount, exist, err := suite.chainB.App.AggregateKeeper.QueryERC20Trace(
+// 		suite.chainB.GetContext(),
+// 		chainBERC20,
+// 		suite.chainA.ChainID,
+// 	)
+// 	suite.Require().NoError(err)
+// 	suite.Require().True(exist)
+// 	suite.Equal(token, strings.ToLower(chainABase.String()))
+// 	suite.Equal(amount.Int64(), int64(0))
 
-	crossChainData := packettypes.CrossChainData{
-		DestChain:       suite.chainB.ChainID,
-		TokenAddress:    chainABase,
-		Receiver:        strings.ToLower(suite.chainB.SenderAddress.String()),
-		Amount:          big.NewInt(100),
-		ContractAddress: "",
-		CallData:        []byte(""),
-		CallbackAddress: common.BigToAddress(big.NewInt(0)),
-		FeeOption:       0,
-	}
-	fee := packettypes.Fee{
-		TokenAddress: chainABase,
-		Amount:       big.NewInt(100),
-	}
+// 	crossChainData := packettypes.CrossChainData{
+// 		DestChain:       suite.chainB.ChainID,
+// 		TokenAddress:    chainABase,
+// 		Receiver:        strings.ToLower(suite.chainB.SenderAddress.String()),
+// 		Amount:          big.NewInt(100),
+// 		ContractAddress: "",
+// 		CallData:        []byte(""),
+// 		CallbackAddress: common.BigToAddress(big.NewInt(0)),
+// 		FeeOption:       0,
+// 	}
+// 	fee := packettypes.Fee{
+// 		TokenAddress: chainABase,
+// 		Amount:       big.NewInt(100),
+// 	}
 
-	balances := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
-	suite.Require().Equal(balances.String(), "100000000000000stake")
-	suite.CrossChainCall(suite.chainA, crossChainData, fee)
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
-	suite.Require().Equal(balances.String(), "99999999999800stake")
+// 	balances := suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
+// 	suite.Require().Equal(balances.String(), "100000000000000stake")
+// 	suite.CrossChainCall(suite.chainA, crossChainData, fee)
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
+// 	suite.Require().Equal(balances.String(), "99999999999800stake")
 
-	crossChainAddr, err := sdk.AccAddressFromHex(hex.EncodeToString(crosschaincontract.CrossChainContractAddress.Bytes()))
-	suite.Require().NoError(err)
-	packetAddr, err := sdk.AccAddressFromHex(hex.EncodeToString(packetcontract.PacketContractAddress.Bytes()))
-	suite.Require().NoError(err)
+// 	crossChainAddr, err := sdk.AccAddressFromHex(hex.EncodeToString(crosschaincontract.CrossChainContractAddress.Bytes()))
+// 	suite.Require().NoError(err)
+// 	packetAddr, err := sdk.AccAddressFromHex(hex.EncodeToString(packetcontract.PacketContractAddress.Bytes()))
+// 	suite.Require().NoError(err)
 
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), crossChainAddr, "stake")
-	suite.Require().Equal(balances.String(), "100stake")
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), packetAddr, "stake")
-	suite.Require().Equal(balances.String(), "100stake")
-	outToken := suite.OutTokens(suite.chainA, chainABase, suite.chainB.ChainID)
-	suite.Require().Equal(int64(100), outToken.Int64())
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), crossChainAddr, "stake")
+// 	suite.Require().Equal(balances.String(), "100stake")
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), packetAddr, "stake")
+// 	suite.Require().Equal(balances.String(), "100stake")
+// 	outToken := suite.OutTokens(suite.chainA, chainABase, suite.chainB.ChainID)
+// 	suite.Require().Equal(int64(100), outToken.Int64())
 
-	// check packet fees
-	fees := suite.GetPacketFees(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
-	suite.Require().Equal(chainABase, fees.TokenAddress)
-	suite.Require().Equal(big.NewInt(100).Int64(), fees.Amount.Int64())
+// 	// check packet fees
+// 	fees := suite.GetPacketFees(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
+// 	suite.Require().Equal(chainABase, fees.TokenAddress)
+// 	suite.Require().Equal(big.NewInt(100).Int64(), fees.Amount.Int64())
 
-	// packet and ack
-	decodeString, err := hex.DecodeString("0000000000000000000000000000000000000000000000000000000000000064")
-	suite.Require().NoError(err)
-	transferData := packettypes.TransferData{
-		Receiver: strings.ToLower(crossChainData.Receiver),
-		Amount:   decodeString,
-		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
-		OriToken: "",
-	}
-	transferDataAbi, err := transferData.ABIPack()
-	suite.Require().NoError(err)
+// 	// packet and ack
+// 	decodeString, err := hex.DecodeString("0000000000000000000000000000000000000000000000000000000000000064")
+// 	suite.Require().NoError(err)
+// 	transferData := packettypes.TransferData{
+// 		Receiver: strings.ToLower(crossChainData.Receiver),
+// 		Amount:   decodeString,
+// 		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
+// 		OriToken: "",
+// 	}
+// 	transferDataAbi, err := transferData.ABIPack()
+// 	suite.Require().NoError(err)
 
-	packet := packettypes.Packet{
-		SourceChain:      suite.chainA.ChainID,
-		DestinationChain: suite.chainB.ChainID,
-		Sequence:         1,
-		Sender:           strings.ToLower(suite.chainA.SenderAddress.String()),
-		TransferData:     transferDataAbi,
-		CallData:         []byte(""),
-		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
-		FeeOption:        0,
-	}
-	ack := packettypes.NewResultAcknowledgement(
-		0,
-		[]byte(""),
-		"",
-		strings.ToLower(suite.chainB.SenderAcc.String()),
-	)
-	ackData, err := ack.ABIPack()
-	suite.Require().NoError(err)
+// 	packet := packettypes.Packet{
+// 		SourceChain:      suite.chainA.ChainID,
+// 		DestinationChain: suite.chainB.ChainID,
+// 		Sequence:         1,
+// 		Sender:           strings.ToLower(suite.chainA.SenderAddress.String()),
+// 		TransferData:     transferDataAbi,
+// 		CallData:         []byte(""),
+// 		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
+// 		FeeOption:        0,
+// 	}
+// 	ack := packettypes.NewAcknowledgement(
+// 		0,
+// 		[]byte(""),
+// 		"",
+// 		strings.ToLower(suite.chainB.SenderAcc.String()),
+// 		0,
+// 	)
+// 	ackData, err := ack.ABIPack()
+// 	suite.Require().NoError(err)
 
-	// relay
-	err = pathAToB.RelayPacket(packet, ackData)
-	suite.Require().NoError(err)
+// 	// relay
+// 	err = pathAToB.RelayPacket(packet, ackData)
+// 	suite.Require().NoError(err)
 
-	// check balance
-	balance := suite.ERC20Balance(suite.chainB, chainBERC20, suite.chainB.SenderAddress)
-	suite.Require().Equal(int64(100), balance.Int64())
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), packetAddr, "stake")
-	suite.Require().Equal(balances.String(), "0stake")
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
-	suite.Require().Equal(balances.String(), "99999999999900stake")
+// 	// check balance
+// 	balance := suite.ERC20Balance(suite.chainB, chainBERC20, suite.chainB.SenderAddress)
+// 	suite.Require().Equal(int64(100), balance.Int64())
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), packetAddr, "stake")
+// 	suite.Require().Equal(balances.String(), "0stake")
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
+// 	suite.Require().Equal(balances.String(), "99999999999900stake")
 
-	// check bindings
-	bindings := suite.Bindings(suite.chainB, chainBERC20, suite.chainA.ChainID)
-	suite.Equal(bindings.OriToken, strings.ToLower(chainABase.String()))
-	suite.Equal(bindings.Amount.Int64(), int64(100))
-	suite.Equal(bindings.Bound, true)
-	suite.Equal(bindings.Scale, uint8(0))
+// 	// check bindings
+// 	bindings := suite.Bindings(suite.chainB, chainBERC20, suite.chainA.ChainID)
+// 	suite.Equal(bindings.OriToken, strings.ToLower(chainABase.String()))
+// 	suite.Equal(bindings.Amount.Int64(), int64(100))
+// 	suite.Equal(bindings.Bound, true)
+// 	suite.Equal(bindings.Scale, uint8(0))
 
-	// check ack
-	status := suite.GetAckStatus(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
-	suite.Require().Equal(status, uint8(1))
-	ack = suite.GetAck(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
-	suite.Require().Equal(uint64(0), ack.Code)
-	suite.Require().Equal([]byte(""), ack.Result)
-	suite.Require().Equal("", ack.Message)
-	suite.Require().Equal(suite.chainB.SenderAcc.String(), ack.Relayer)
+// 	// check ack
+// 	status := suite.GetAckStatus(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
+// 	suite.Require().Equal(status, uint8(1))
+// 	ack = suite.GetAck(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
+// 	suite.Require().Equal(uint64(0), ack.Code)
+// 	suite.Require().Equal([]byte(""), ack.Result)
+// 	suite.Require().Equal("", ack.Message)
+// 	suite.Require().Equal(suite.chainB.SenderAcc.String(), ack.Relayer)
 
-	crossChainData = packettypes.CrossChainData{
-		DestChain:       suite.chainA.ChainID,
-		TokenAddress:    chainBERC20,
-		Receiver:        strings.ToLower(suite.chainA.SenderAddress.String()),
-		Amount:          big.NewInt(100),
-		ContractAddress: "",
-		CallData:        []byte(""),
-		CallbackAddress: common.BigToAddress(big.NewInt(0)),
-		FeeOption:       0,
-	}
-	fee = packettypes.Fee{
-		TokenAddress: chainBERC20,
-		Amount:       big.NewInt(0),
-	}
-	suite.Approve(suite.chainB, chainBERC20, crosschaincontract.CrossChainContractAddress, big.NewInt(100))
-	suite.CrossChainCall(suite.chainB, crossChainData, fee)
-	balance = suite.ERC20Balance(suite.chainB, chainBERC20, suite.chainB.SenderAddress)
-	suite.Require().Equal(int64(0), balance.Int64())
+// 	crossChainData = packettypes.CrossChainData{
+// 		DestChain:       suite.chainA.ChainID,
+// 		TokenAddress:    chainBERC20,
+// 		Receiver:        strings.ToLower(suite.chainA.SenderAddress.String()),
+// 		Amount:          big.NewInt(100),
+// 		ContractAddress: "",
+// 		CallData:        []byte(""),
+// 		CallbackAddress: common.BigToAddress(big.NewInt(0)),
+// 		FeeOption:       0,
+// 	}
+// 	fee = packettypes.Fee{
+// 		TokenAddress: chainBERC20,
+// 		Amount:       big.NewInt(0),
+// 	}
+// 	suite.Approve(suite.chainB, chainBERC20, crosschaincontract.CrossChainContractAddress, big.NewInt(100))
+// 	suite.CrossChainCall(suite.chainB, crossChainData, fee)
+// 	balance = suite.ERC20Balance(suite.chainB, chainBERC20, suite.chainB.SenderAddress)
+// 	suite.Require().Equal(int64(0), balance.Int64())
 
-	// check bindings
-	bindings = suite.Bindings(suite.chainB, chainBERC20, suite.chainA.ChainID)
-	suite.Equal(bindings.OriToken, strings.ToLower(chainABase.String()))
-	suite.Equal(bindings.Amount.Int64(), int64(0))
-	suite.Equal(bindings.Bound, true)
-	suite.Equal(bindings.Scale, uint8(0))
+// 	// check bindings
+// 	bindings = suite.Bindings(suite.chainB, chainBERC20, suite.chainA.ChainID)
+// 	suite.Equal(bindings.OriToken, strings.ToLower(chainABase.String()))
+// 	suite.Equal(bindings.Amount.Int64(), int64(0))
+// 	suite.Equal(bindings.Bound, true)
+// 	suite.Equal(bindings.Scale, uint8(0))
 
-	// check packet fees
-	fees = suite.GetPacketFees(suite.chainB, suite.chainB.ChainID, suite.chainA.ChainID, 1)
-	suite.Require().Equal(chainBERC20, fees.TokenAddress)
-	suite.Require().Equal(big.NewInt(0).Int64(), fees.Amount.Int64())
+// 	// check packet fees
+// 	fees = suite.GetPacketFees(suite.chainB, suite.chainB.ChainID, suite.chainA.ChainID, 1)
+// 	suite.Require().Equal(chainBERC20, fees.TokenAddress)
+// 	suite.Require().Equal(big.NewInt(0).Int64(), fees.Amount.Int64())
 
-	// packet and ack
-	decodeString, err = hex.DecodeString("0000000000000000000000000000000000000000000000000000000000000064")
-	suite.Require().NoError(err)
-	transferData = packettypes.TransferData{
-		Receiver: strings.ToLower(crossChainData.Receiver),
-		Amount:   decodeString,
-		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
-		OriToken: strings.ToLower(chainABase.String()),
-	}
-	transferDataAbi, err = transferData.ABIPack()
-	suite.Require().NoError(err)
+// 	// packet and ack
+// 	decodeString, err = hex.DecodeString("0000000000000000000000000000000000000000000000000000000000000064")
+// 	suite.Require().NoError(err)
+// 	transferData = packettypes.TransferData{
+// 		Receiver: strings.ToLower(crossChainData.Receiver),
+// 		Amount:   decodeString,
+// 		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
+// 		OriToken: strings.ToLower(chainABase.String()),
+// 	}
+// 	transferDataAbi, err = transferData.ABIPack()
+// 	suite.Require().NoError(err)
 
-	packet = packettypes.Packet{
-		SourceChain:      suite.chainB.ChainID,
-		DestinationChain: suite.chainA.ChainID,
-		Sequence:         1,
-		Sender:           strings.ToLower(suite.chainB.SenderAddress.String()),
-		TransferData:     transferDataAbi,
-		CallData:         []byte(""),
-		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
-		FeeOption:        0,
-	}
-	ack = packettypes.NewResultAcknowledgement(
-		0,
-		[]byte(""),
-		"",
-		strings.ToLower(suite.chainB.SenderAcc.String()),
-	)
-	ackData, err = ack.ABIPack()
-	suite.Require().NoError(err)
+// 	packet = packettypes.Packet{
+// 		SourceChain:      suite.chainB.ChainID,
+// 		DestinationChain: suite.chainA.ChainID,
+// 		Sequence:         1,
+// 		Sender:           strings.ToLower(suite.chainB.SenderAddress.String()),
+// 		TransferData:     transferDataAbi,
+// 		CallData:         []byte(""),
+// 		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
+// 		FeeOption:        0,
+// 	}
+// 	ack = packettypes.NewAcknowledgement(
+// 		0,
+// 		[]byte(""),
+// 		"",
+// 		strings.ToLower(suite.chainB.SenderAcc.String()),
+// 		0,
+// 	)
+// 	ackData, err = ack.ABIPack()
+// 	suite.Require().NoError(err)
 
-	// relay
-	err = pathAToB.RelayPacket(packet, ackData)
-	suite.Require().NoError(err)
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
-	suite.Require().Equal(balances.String(), "100000000000000stake")
+// 	// relay
+// 	err = pathAToB.RelayPacket(packet, ackData)
+// 	suite.Require().NoError(err)
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), suite.chainA.SenderAcc, "stake")
+// 	suite.Require().Equal(balances.String(), "100000000000000stake")
 
-	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), crossChainAddr, "stake")
-	suite.Require().Equal(balances.String(), "0stake")
-	outToken = suite.OutTokens(suite.chainA, chainABase, suite.chainB.ChainID)
-	suite.Require().Equal(int64(0), outToken.Int64())
-}
+// 	balances = suite.chainA.App.BankKeeper.GetBalance(suite.chainA.GetContext(), crossChainAddr, "stake")
+// 	suite.Require().Equal(balances.String(), "0stake")
+// 	outToken = suite.OutTokens(suite.chainA, chainABase, suite.chainB.ChainID)
+// 	suite.Require().Equal(int64(0), outToken.Int64())
+// }
 
-func (suite *XIBCTestSuite) TestCrossChainCallAgent() {
-	suite.SetupTest()
+// func (suite *XIBCTestSuite) TestCrossChainCallAgent() {
+// 	suite.SetupTest()
 
-	// setup testing conditions
-	pathAToB := xibctesting.NewPath(suite.chainA, suite.chainB)
-	pathBToC := xibctesting.NewPath(suite.chainA, suite.chainB)
+// 	// setup testing conditions
+// 	pathAToB := xibctesting.NewPath(suite.chainA, suite.chainB)
+// 	pathBToC := xibctesting.NewPath(suite.chainA, suite.chainB)
 
-	suite.coordinator.SetupClients(pathAToB)
-	suite.coordinator.SetupClients(pathBToC)
+// 	suite.coordinator.SetupClients(pathAToB)
+// 	suite.coordinator.SetupClients(pathBToC)
 
-	// deploy ERC20
-	chainAERC20 := suite.DeployERC20ByCrossChain(suite.chainA)
-	chainBERC20 := suite.DeployERC20ByCrossChain(suite.chainB)
-	chainCERC20 := suite.DeployERC20ByCrossChain(suite.chainC)
+// 	// deploy ERC20
+// 	chainAERC20 := suite.DeployERC20ByCrossChain(suite.chainA)
+// 	chainBERC20 := suite.DeployERC20ByCrossChain(suite.chainB)
+// 	chainCERC20 := suite.DeployERC20ByCrossChain(suite.chainC)
 
-	suite.GrantERC20MintRoleByCrossChain(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
-	suite.MintERC20Token(suite.chainA, suite.chainA.SenderAddress, chainAERC20, big.NewInt(10000))
-	balance := suite.ERC20Balance(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
-	suite.Equal(balance.Int64(), int64(10000))
+// 	suite.GrantERC20MintRoleByCrossChain(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
+// 	suite.MintERC20Token(suite.chainA, suite.chainA.SenderAddress, chainAERC20, big.NewInt(10000))
+// 	balance := suite.ERC20Balance(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
+// 	suite.Equal(balance.Int64(), int64(10000))
 
-	// register erc20 trace A -> B
-	err := suite.chainB.App.AggregateKeeper.RegisterERC20Trace(
-		suite.chainB.GetContext(),
-		chainBERC20,
-		strings.ToLower(chainAERC20.String()),
-		suite.chainA.ChainID,
-		uint8(0),
-	)
-	suite.Require().NoError(err)
-	// check ERC20 trace
-	token, amount, exist, err := suite.chainB.App.AggregateKeeper.QueryERC20Trace(
-		suite.chainB.GetContext(),
-		chainBERC20,
-		suite.chainA.ChainID,
-	)
-	suite.Require().NoError(err)
-	suite.Require().True(exist)
-	suite.Equal(token, strings.ToLower(chainAERC20.String()))
-	suite.Equal(amount.Int64(), int64(0))
+// 	// register erc20 trace A -> B
+// 	err := suite.chainB.App.AggregateKeeper.RegisterERC20Trace(
+// 		suite.chainB.GetContext(),
+// 		chainBERC20,
+// 		strings.ToLower(chainAERC20.String()),
+// 		suite.chainA.ChainID,
+// 		uint8(0),
+// 	)
+// 	suite.Require().NoError(err)
+// 	// check ERC20 trace
+// 	token, amount, exist, err := suite.chainB.App.AggregateKeeper.QueryERC20Trace(
+// 		suite.chainB.GetContext(),
+// 		chainBERC20,
+// 		suite.chainA.ChainID,
+// 	)
+// 	suite.Require().NoError(err)
+// 	suite.Require().True(exist)
+// 	suite.Equal(token, strings.ToLower(chainAERC20.String()))
+// 	suite.Equal(amount.Int64(), int64(0))
 
-	// register erc20 trace B -> C
-	err = suite.chainC.App.AggregateKeeper.RegisterERC20Trace(
-		suite.chainC.GetContext(),
-		chainCERC20,
-		strings.ToLower(chainBERC20.String()),
-		suite.chainB.ChainID,
-		uint8(0),
-	)
-	suite.Require().NoError(err)
-	// check ERC20 trace
-	token, amount, exist, err = suite.chainC.App.AggregateKeeper.QueryERC20Trace(
-		suite.chainC.GetContext(),
-		chainCERC20,
-		suite.chainB.ChainID,
-	)
-	suite.Require().NoError(err)
-	suite.Require().True(exist)
-	suite.Equal(token, strings.ToLower(chainBERC20.String()))
-	suite.Equal(amount.Int64(), int64(0))
+// 	// register erc20 trace B -> C
+// 	err = suite.chainC.App.AggregateKeeper.RegisterERC20Trace(
+// 		suite.chainC.GetContext(),
+// 		chainCERC20,
+// 		strings.ToLower(chainBERC20.String()),
+// 		suite.chainB.ChainID,
+// 		uint8(0),
+// 	)
+// 	suite.Require().NoError(err)
+// 	// check ERC20 trace
+// 	token, amount, exist, err = suite.chainC.App.AggregateKeeper.QueryERC20Trace(
+// 		suite.chainC.GetContext(),
+// 		chainCERC20,
+// 		suite.chainB.ChainID,
+// 	)
+// 	suite.Require().NoError(err)
+// 	suite.Require().True(exist)
+// 	suite.Equal(token, strings.ToLower(chainBERC20.String()))
+// 	suite.Equal(amount.Int64(), int64(0))
 
-	callData, err := agent.AgentContract.ABI.Pack(
-		"send",
-		chainBERC20,
-		suite.chainA.SenderAddress,
-		strings.ToLower(suite.chainA.SenderAddress.String()),
-		suite.chainC.ChainID,
-		big.NewInt(1000),
-	)
-	suite.Require().NoError(err)
+// 	callData, err := agentcontract.AgentContract.ABI.Pack(
+// 		"send",
+// 		chainBERC20,
+// 		suite.chainA.SenderAddress,
+// 		strings.ToLower(suite.chainA.SenderAddress.String()),
+// 		suite.chainC.ChainID,
+// 		big.NewInt(1000),
+// 	)
+// 	suite.Require().NoError(err)
 
-	crossChainData := packettypes.CrossChainData{
-		DestChain:       suite.chainB.ChainID,
-		TokenAddress:    chainAERC20,
-		Receiver:        strings.ToLower(agent.AgentContractAddress.String()),
-		Amount:          big.NewInt(2000),
-		ContractAddress: syscontracts.AgentContractAddress,
-		CallData:        callData,
-		CallbackAddress: common.BigToAddress(big.NewInt(0)),
-		FeeOption:       0,
-	}
-	fee := packettypes.Fee{
-		TokenAddress: chainAERC20,
-		Amount:       big.NewInt(0),
-	}
+// 	crossChainData := packettypes.CrossChainData{
+// 		DestChain:       suite.chainB.ChainID,
+// 		TokenAddress:    chainAERC20,
+// 		Receiver:        strings.ToLower(agentcontract.AgentContractAddress.String()),
+// 		Amount:          big.NewInt(2000),
+// 		ContractAddress: syscontracts.AgentContractAddress,
+// 		CallData:        callData,
+// 		CallbackAddress: common.BigToAddress(big.NewInt(0)),
+// 		FeeOption:       0,
+// 	}
+// 	fee := packettypes.Fee{
+// 		TokenAddress: chainAERC20,
+// 		Amount:       big.NewInt(0),
+// 	}
 
-	// send CrossChainCall Tx
-	suite.Approve(suite.chainA, chainAERC20, crosschaincontract.CrossChainContractAddress, big.NewInt(2000))
-	suite.CrossChainCall(suite.chainA, crossChainData, fee)
-	balance = suite.ERC20Balance(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
-	suite.Equal(balance.Int64(), int64(8000))
-	balance = suite.ERC20Balance(suite.chainA, chainAERC20, crosschaincontract.CrossChainContractAddress)
-	suite.Equal(balance.Int64(), int64(2000))
-	balance = suite.ERC20Balance(suite.chainA, chainAERC20, packetcontract.PacketContractAddress)
-	suite.Equal(balance.Int64(), int64(0))
-	outToken := suite.OutTokens(suite.chainA, chainAERC20, suite.chainB.ChainID)
-	suite.Require().Equal(int64(2000), outToken.Int64())
+// 	// send CrossChainCall Tx
+// 	suite.Approve(suite.chainA, chainAERC20, crosschaincontract.CrossChainContractAddress, big.NewInt(2000))
+// 	suite.CrossChainCall(suite.chainA, crossChainData, fee)
+// 	balance = suite.ERC20Balance(suite.chainA, chainAERC20, suite.chainA.SenderAddress)
+// 	suite.Equal(balance.Int64(), int64(8000))
+// 	balance = suite.ERC20Balance(suite.chainA, chainAERC20, crosschaincontract.CrossChainContractAddress)
+// 	suite.Equal(balance.Int64(), int64(2000))
+// 	balance = suite.ERC20Balance(suite.chainA, chainAERC20, packetcontract.PacketContractAddress)
+// 	suite.Equal(balance.Int64(), int64(0))
+// 	outToken := suite.OutTokens(suite.chainA, chainAERC20, suite.chainB.ChainID)
+// 	suite.Require().Equal(int64(2000), outToken.Int64())
 
-	// check packet fees
-	fees := suite.GetPacketFees(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
-	suite.Require().Equal(chainAERC20, fees.TokenAddress)
-	suite.Require().Equal(big.NewInt(0).Int64(), fees.Amount.Int64())
+// 	// check packet fees
+// 	fees := suite.GetPacketFees(suite.chainA, suite.chainA.ChainID, suite.chainB.ChainID, 1)
+// 	suite.Require().Equal(chainAERC20, fees.TokenAddress)
+// 	suite.Require().Equal(big.NewInt(0).Int64(), fees.Amount.Int64())
 
-	// packet and ack
-	hexAmount, err := hex.DecodeString("00000000000000000000000000000000000000000000000000000000000007d0")
-	suite.Require().NoError(err)
-	transferData := packettypes.TransferData{
-		Receiver: strings.ToLower(crossChainData.Receiver),
-		Amount:   hexAmount,
-		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
-		OriToken: "",
-	}
-	transferDataAbi, err := transferData.ABIPack()
-	suite.Require().NoError(err)
+// 	// packet and ack
+// 	hexAmount, err := hex.DecodeString("00000000000000000000000000000000000000000000000000000000000007d0")
+// 	suite.Require().NoError(err)
+// 	transferData := packettypes.TransferData{
+// 		Receiver: strings.ToLower(crossChainData.Receiver),
+// 		Amount:   hexAmount,
+// 		Token:    strings.ToLower(crossChainData.TokenAddress.String()),
+// 		OriToken: "",
+// 	}
+// 	transferDataAbi, err := transferData.ABIPack()
+// 	suite.Require().NoError(err)
 
-	packetCallData := &packettypes.CallData{
-		ContractAddress: syscontracts.AgentContractAddress,
-		CallData:        callData,
-	}
-	callDataAbi, err := packetCallData.ABIPack()
-	suite.Require().NoError(err)
+// 	packetCallData := &packettypes.CallData{
+// 		ContractAddress: syscontracts.AgentContractAddress,
+// 		CallData:        callData,
+// 	}
+// 	callDataAbi, err := packetCallData.ABIPack()
+// 	suite.Require().NoError(err)
 
-	packet := packettypes.Packet{
-		SourceChain:      suite.chainA.ChainID,
-		DestinationChain: suite.chainB.ChainID,
-		Sequence:         1,
-		Sender:           strings.ToLower(suite.chainA.SenderAddress.String()),
-		TransferData:     transferDataAbi,
-		CallData:         callDataAbi,
-		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
-		FeeOption:        0,
-	}
-	ack := packettypes.NewResultAcknowledgement(
-		0,
-		[]byte(""),
-		"",
-		strings.ToLower(suite.chainB.SenderAcc.String()),
-	)
-	fmt.Println(strings.ToLower(suite.chainB.SenderAcc.String()))
-	ackData, err := ack.ABIPack()
-	suite.Require().NoError(err)
-	suite.Require().NotNil(ackData)
-	suite.Require().NotNil(packet)
+// 	packet := packettypes.Packet{
+// 		SourceChain:      suite.chainA.ChainID,
+// 		DestinationChain: suite.chainB.ChainID,
+// 		Sequence:         1,
+// 		Sender:           strings.ToLower(suite.chainA.SenderAddress.String()),
+// 		TransferData:     transferDataAbi,
+// 		CallData:         callDataAbi,
+// 		CallbackAddress:  common.BigToAddress(big.NewInt(0)).String(),
+// 		FeeOption:        0,
+// 	}
+// 	ack := packettypes.NewAcknowledgement(
+// 		0,
+// 		[]byte(""),
+// 		"",
+// 		strings.ToLower(suite.chainB.SenderAcc.String()),
+// 		0,
+// 	)
+// 	fmt.Println(strings.ToLower(suite.chainB.SenderAcc.String()))
+// 	ackData, err := ack.ABIPack()
+// 	suite.Require().NoError(err)
+// 	suite.Require().NotNil(ackData)
+// 	suite.Require().NotNil(packet)
 
-	// relay
-	//err = pathAToB.RelayPacket(packet, ackData)
-	//suite.Require().NoError(err)
-}
+// 	// relay
+// 	//err = pathAToB.RelayPacket(packet, ackData)
+// 	//suite.Require().NoError(err)
+// }
 
 // ================================================================================================================
 // CrossChain functions
@@ -597,7 +631,7 @@ func (suite *XIBCTestSuite) Bindings(
 // Packet functions
 // ================================================================================================================
 func (suite *XIBCTestSuite) GetPacketFees(fromChain *xibctesting.TestChain, srcChain, destChain string, sequence uint64) packettypes.Fee {
-	data := []byte(srcChain + "/" + destChain + "/" + strconv.FormatUint(sequence, 10))
+	data := []byte(destChain + "/" + strconv.FormatUint(sequence, 10))
 	packet := packetcontract.PacketContract.ABI
 
 	res, err := fromChain.App.AggregateKeeper.CallEVM(
